+++ src/lmdesktopplus/__init__.py
@@ -1,3 +1,3 @@
 """LMDesktopPlus machine UI."""
 
-__version__ = "0.4.6"
+__version__ = "0.5.0"